refactor: Utils Folder Structure
